+++ NEWS.md
@@ -1,7 +1,7 @@
 # satres (development version)
 * Functions to transform an object: `clip_bands()`.
 * Functions to query an object: `get_band_names()`, `get_spectral_band_names()`.
-* Change `satres()` parameter `only_bands` to `only_spectral` to avoid ambiguity.
+* Change `satres()` parameter `only_bands` to `only_spectral_bands` to avoid ambiguity.
 
 # satres 1.0.0
 * Functionality for creating `satres` objects and exporting them.